+++ src/lib.rs
@@ -87,7 +87,6 @@ mod macros;
 pub mod measurement;
 mod plot;
 pub mod profiler;
-mod program;
 mod report;
 mod routine;
 mod stats;
@@ -98,7 +97,6 @@ use std::default::Default;
 use std::fmt;
 use std::iter::IntoIterator;
 use std::marker::PhantomData;
-use std::process::Command;
 use std::time::Duration;
 
 use crate::benchmark::BenchmarkConfig;
@@ -712,93 +710,6 @@ impl Default for Criterion {
             measurement: WallTime,
             profiler: Box::new(RefCell::new(ExternalProfiler)),
         }
-    }
-}
-
-impl Criterion<WallTime> {
-    /// Benchmarks an external program
-    ///
-    /// The external program must:
-    ///
-    /// * Read the number of iterations from stdin
-    /// * Execute the routine to benchmark that many times
-    /// * Print the elapsed time (in nanoseconds) to stdout
-    ///
-    /// ```rust,no_run
-    /// # use std::io::{self, BufRead};
-    /// # use std::time::Instant;
-    /// # use std::time::Duration;
-    /// # trait DurationExt { fn to_nanos(&self) -> u64 { 0 } }
-    /// # impl DurationExt for Duration {}
-    /// // Example of an external program that implements this protocol
-    ///
-    /// fn main() {
-    ///     let stdin = io::stdin();
-    ///     let ref mut stdin = stdin.lock();
-    ///
-    ///     // For each line in stdin
-    ///     for line in stdin.lines() {
-    ///         // Parse line as the number of iterations
-    ///         let iters: u64 = line.unwrap().trim().parse().unwrap();
-    ///
-    ///         // Setup
-    ///
-    ///         // Benchmark
-    ///         let start = Instant::now();
-    ///         // Execute the routine "iters" times
-    ///         for _ in 0..iters {
-    ///             // Code to benchmark goes here
-    ///         }
-    ///         let elapsed = start.elapsed();
-    ///
-    ///         // Teardown
-    ///
-    ///         // Report elapsed time in nanoseconds to stdout
-    ///         println!("{}", elapsed.to_nanos());
-    ///     }
-    /// }
-    /// ```
-    #[deprecated(
-        since = "0.2.6",
-        note = "External program benchmarks were rarely used and are awkward to maintain, so they are scheduled for deletion in 0.3.0"
-    )]
-    #[allow(deprecated)]
-    pub fn bench_program(&mut self, id: &str, program: Command) -> &mut Criterion<WallTime> {
-        self.bench(id, Benchmark::new_external(id, program))
-    }
-
-    /// Benchmarks an external program under various inputs
-    ///
-    /// This is a convenience method to execute several related benchmarks. Each benchmark will
-    /// receive the id: `${id}/${input}`.
-    #[deprecated(
-        since = "0.2.6",
-        note = "External program benchmarks were rarely used and are awkward to maintain, so they are scheduled for deletion in 0.3.0"
-    )]
-    #[allow(deprecated)]
-    pub fn bench_program_over_inputs<I, F>(
-        &mut self,
-        id: &str,
-        mut program: F,
-        inputs: I,
-    ) -> &mut Criterion<WallTime>
-    where
-        F: FnMut() -> Command + 'static,
-        I: IntoIterator,
-        I::Item: fmt::Debug + 'static,
-    {
-        self.bench(
-            id,
-            ParameterizedBenchmark::new_external(
-                id,
-                move |i| {
-                    let mut command = program();
-                    command.arg(format!("{:?}", i));
-                    command
-                },
-                inputs,
-            ),
-        )
     }
 }